+++ meta_auth/codenames/screening.py
@@ -1,5 +1,8 @@
-from edc_auth.codenames import screening
+from copy import copy
+from edc_auth.codenames import screening as default
+from sarscov2.auth import sarscov2_codenames
 
+screening = copy(default)
 screening += [
     "meta_screening.add_screeningpartone",
     "meta_screening.add_screeningpartthree",
@@ -27,3 +30,5 @@ screening += [
     "meta_screening.view_subjectscreening",
     "meta_screening.view_subjectrefusal",
 ]
+screening.extend(sarscov2_codenames)
+screening.sort()