+++ meta_visit_schedule/visit_schedules/crfs.py
@@ -4,27 +4,25 @@ crfs_prn = FormsCollection(
     Crf(show_order=10, model="meta_subject.bloodresultsfbc"),
     Crf(show_order=20, model="meta_subject.bloodresultsglu"),
     Crf(show_order=30, model="meta_subject.bloodresultshba1c"),
-    Crf(show_order=35, model="meta_subject.bloodresultslipid"),
     Crf(show_order=40, model="meta_subject.bloodresultslft"),
-    Crf(show_order=50, model="meta_subject.bloodresultsrft"),
-    Crf(show_order=55, model="meta_subject.malariatest"),
-    Crf(show_order=60, model="meta_subject.urinedipsticktest"),
+    Crf(show_order=50, model="meta_subject.bloodresultslipid"),
+    Crf(show_order=60, model="meta_subject.bloodresultsrft"),
+    Crf(show_order=70, model="meta_subject.healtheconomics"),
+    Crf(show_order=80, model="meta_subject.malariatest"),
+    Crf(show_order=90, model="meta_subject.urinedipsticktest"),
     name="prn",
 )
 
 crfs_unscheduled = FormsCollection(
     Crf(show_order=10, model="meta_subject.followupvitals"),
     Crf(show_order=20, model="meta_subject.followup"),
-    Crf(show_order=30, model="meta_subject.bloodresultsglu"),
-    Crf(show_order=40, model="meta_subject.bloodresultsfbc"),
-    Crf(show_order=50, model="meta_subject.bloodresultslipid"),
-    Crf(show_order=60, model="meta_subject.bloodresultslft"),
-    Crf(show_order=70, model="meta_subject.bloodresultsrft"),
-    Crf(show_order=80, model="meta_subject.malariatest"),
-    Crf(show_order=90, model="meta_subject.urinedipsticktest"),
+    Crf(show_order=30, model="meta_subject.medicationadherence"),
     name="unscheduled",
 )
 
+crfs_missed = FormsCollection(
+    Crf(show_order=10, model="meta_subject.missedvisit"), name="missed",
+)
 
 crfs_d1 = FormsCollection(
     Crf(show_order=10, model="meta_subject.physicalexam"),
@@ -66,14 +64,11 @@ crfs_6m = FormsCollection(
     Crf(show_order=10, model="meta_subject.followupvitals"),
     Crf(show_order=15, model="meta_subject.followup"),
     Crf(show_order=20, model="meta_subject.medicationadherence"),
-    Crf(show_order=30, model="meta_subject.bloodresultsglu"),
+    Crf(show_order=25, model="meta_subject.glucose"),
     Crf(show_order=40, model="meta_subject.bloodresultshba1c"),
     Crf(show_order=50, model="meta_subject.bloodresultsfbc"),
-    Crf(show_order=60, model="meta_subject.bloodresultslipid"),
     Crf(show_order=70, model="meta_subject.bloodresultslft"),
     Crf(show_order=80, model="meta_subject.bloodresultsrft"),
-    Crf(show_order=90, model="meta_subject.malariatest"),
-    Crf(show_order=100, model="meta_subject.urinedipsticktest"),
     name="6m",
 )
 
@@ -90,13 +85,11 @@ crfs_12m = FormsCollection(
     Crf(show_order=10, model="meta_subject.followupvitals"),
     Crf(show_order=15, model="meta_subject.followup"),
     Crf(show_order=20, model="meta_subject.medicationadherence"),
-    Crf(show_order=30, model="meta_subject.bloodresultsglu"),
+    Crf(show_order=30, model="meta_subject.glucose"),
     Crf(show_order=40, model="meta_subject.bloodresultshba1c"),
     Crf(show_order=50, model="meta_subject.bloodresultsfbc"),
     Crf(show_order=55, model="meta_subject.bloodresultslipid"),
     Crf(show_order=60, model="meta_subject.bloodresultslft"),
     Crf(show_order=70, model="meta_subject.bloodresultsrft"),
-    Crf(show_order=80, model="meta_subject.malariatest"),
-    Crf(show_order=90, model="meta_subject.urinedipsticktest"),
     name="12m",
 )